testing/web-platform/tests/mediacapture-streams/MediaStreamTrack-MediaElement-disabled-video-is-black.https.html

<!doctype html>
<html>
<head>
<title>A disabled video track is rendered as blackness</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#introduction">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#mediastreams-as-media-elements">
</head>
<body>
<p class="instructions">When prompted, accept to share your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that a disabled video track in a
MediaStream is rendered as blackness.</p>
<video id="vid"></video>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=permission-helper.js></script>
<script>
const vid = document.getElementById("vid");
const cv = document.createElement("canvas");
promise_test(async t => {
  await setMediaPermission("granted", ["camera"]);
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  t.add_cleanup(() => {
    for (let track of stream.getTracks()) {
      track.stop();
    }
    vid.srcObject = null;
  });

  stream.getTracks()[0].enabled = false;
  vid.srcObject = stream;
  vid.play();
  await new Promise(r => vid.onloadeddata = r);

  cv.width = vid.offsetWidth;
  cv.height = vid.offsetHeight;
  const ctx = cv.getContext("2d");
  ctx.drawImage(vid,0,0);
  const imageData = ctx.getImageData(0, 0, cv.width, cv.height);
  for (let i = 0; i < imageData.data.length; i+=4) {
    assert_equals(imageData.data[i], 0, "No red component in pixel #" + i);
    assert_equals(imageData.data[i + 1], 0, "No green component in pixel #" + i);
    assert_equals(imageData.data[i + 2], 0, "No blue component in pixel #" + i);
    assert_equals(imageData.data[i + 3], 255, "No transparency in pixel #" + i);
  }
}, "Tests that a disabled video track in a MediaStream is rendered as blackness");
</script>
</body>
</html>
